Move Azure-specific test fixture to its own module
