+++ extensions/azure/azure-test/build.gradle.kts
@@ -19,12 +19,12 @@ plugins {
 }
 
 val awsVersion: String by project
+val storageBlobVersion: String by project;
 
 dependencies {
     api(project(":spi"))
 
-    testFixturesApi(platform("com.amazonaws:aws-java-sdk-bom:1.11.1018"))
-    testFixturesApi("com.amazonaws:aws-java-sdk-s3")
+    testFixturesApi("com.azure:azure-storage-blob:${storageBlobVersion}")
 
     testFixturesImplementation("org.junit.jupiter:junit-jupiter-api:5.5.2")
     testFixturesRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.5.2")
@@ -33,8 +33,8 @@ dependencies {
 
 publishing {
     publications {
-        create<MavenPublication>("aws.test") {
-            artifactId = "aws.test"
+        create<MavenPublication>("azure.test") {
+            artifactId = "azure.test"
             from(components["java"])
         }
     }
