+++ extensions/azure/blob/provision/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
 
     implementation("com.azure:azure-storage-blob:${storageBlobVersion}")
 
+    testImplementation(testFixtures(project(":extensions:azure:azure-test")))
     testImplementation(testFixtures(project(":common:util")))
 
 }
