+++ settings.gradle.kts
@@ -40,6 +40,7 @@ include(":extensions:aws:aws-test")
 include(":extensions:azure:blob:blob-schema")
 include(":extensions:azure:blob:provision")
 include(":extensions:azure:events")
+include(":extensions:azure:azure-test")
 include(":extensions:azure:transfer-process-store-cosmos")
 include(":extensions:azure:vault")
 include(":extensions:azure:blob:api")
